Fix type bugs with NumPy v2 (#1994)
After NumPy v2 the NumPy precision is preferred over the precision
provided by any native Python types involved in calculations. This
causes the result types of arithmetic operators to change. This was not
a problem in the devel branch due to our use of `np.result_type` to
calculate the types using information about both the type and the
precision.

This PR fixes the problem to prepare the patch release. It does this by
ignoring native types. The fix handles most cases but is not perfect. It
is very difficult to do better without making the changes in #1756.

+++ tests/epyccel/recognised_functions/test_numpy_funcs.py
@@ -5388,10 +5388,10 @@ def test_numpy_linspace_scalar(language):
 
     @template('T', ['int', 'int8', 'int16', 'int32', 'int64', 'float', 'float32', 'float64'])
     def get_linspace(start : 'T', steps : int, num : int):
-        from numpy import linspace
+        from numpy import linspace, float64
         stop = start + steps
         b = linspace(start, stop, num)
-        x = 0.0
+        x = float64(0.0)
         for bi in b:
             x += bi
         return x
